added InitJython so we can define the jar start point if we want to
useful info:
in the pom file we're currently delegating the start point to the one
defined in jython-standalone. in jars, this start point has to be
defined in the META-INF which describes which class to load and how to
do it. therefore, if we create our own InitJython we can remove the
shading of the default jython-standalone start point and use our own
instead which will allow us to customize the initialization to the way
we prefer :)